+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Please add your own contribution below inside the Master section
 Bug-fixes within the same version aren't needed
 
 ## Master
+* fix debug problem for windows users (#694) - @connectdotz
 -->
 
 ### 4.0.1